feat(tarko-ui): integrate @assistant-ui/react-markdown for smooth streaming
- Replace custom smooth streaming with @assistant-ui/react-markdown
- Add StandaloneMarkdownText component for independent usage
- Update SmoothMarkdownRenderer to use assistant-ui styling
- Maintain backward compatibility with existing MarkdownRenderer API
- Import assistant-ui dot.css styles for smooth rendering effects
