+++ multimodal/tarko/ui/src/components/markdown-renderer/assistant-ui/MarkdownText.tsx
@@ -1,13 +1,6 @@
 import React, { memo } from 'react';
-import ReactMarkdown from 'react-markdown';
-import remarkGfm from 'remark-gfm';
-import remarkMath from 'remark-math';
-import rehypeKatex from 'rehype-katex';
-import { remarkAlert } from 'remark-github-blockquote-alert';
-import { useMarkdownComponents } from '../components';
-
-// Fallback to regular react-markdown for now
-// TODO: Implement proper assistant-ui integration when the API is more stable
+import { StandaloneMarkdownText } from './StandaloneMarkdownText';
+import '@assistant-ui/react-markdown/styles/dot.css';
 
 interface MarkdownTextProps {
   content: string;
@@ -22,22 +15,16 @@ const MarkdownTextImpl: React.FC<MarkdownTextProps> = ({
   onImageClick,
   codeBlockStyle,
   className = '',
-  smooth = true, // Note: smooth is not yet implemented in this fallback
+  smooth = true,
 }) => {
-  const components = useMarkdownComponents({
-    onImageClick: onImageClick || (() => {}),
-    codeBlockStyle,
-  });
-
   return (
     <div className={className}>
-      <ReactMarkdown
-        remarkPlugins={[remarkGfm, remarkMath, remarkAlert]}
-        rehypePlugins={[rehypeKatex]}
-        components={components}
-      >
-        {content}
-      </ReactMarkdown>
+      <StandaloneMarkdownText
+        content={content}
+        onImageClick={onImageClick}
+        codeBlockStyle={codeBlockStyle}
+        smooth={smooth}
+      />
     </div>
   );
 };
