+++ multimodal/tarko/ui/src/components/markdown-renderer/index.tsx
@@ -1,3 +1,5 @@
 export { MarkdownRenderer } from './MarkdownRenderer';
+export { MarkdownText } from './assistant-ui/MarkdownText';
+export { StandaloneMarkdownText } from './assistant-ui/StandaloneMarkdownText';
 export * from './components';
 export * from './utils';
